backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
